refactor: 핸들러, 초기로직 분리
